Add: Find a composer Package for post meta (YAML FRONT MATTER)
The url package that has been used in PHP is YAML FRONT MATTER

https://github.com/spatie/yaml-front-matter

This way, we can create a more "dynamic" website based on the meta data
post we have on each html site.

Currently, this meta data is of course hard coded. In the future, we will
make sure to save those data into the database, so it will go faster